Refactor ptrace functionality into files separate from os.c
