+++ core/CMakeLists.txt
@@ -398,6 +398,9 @@ if (UNIX)
       if (AARCH64 OR (X86 AND X64))
         set(OS_SRCS ${OS_SRCS} unix/coredump.c)
       endif ()
+      if (AARCH64 OR (X86 AND X64))
+        set(OS_SRCS ${OS_SRCS} unix/ptrace_lib.c unix/ptrace_attach.c)
+      endif ()
     endif ()
     set(OS_SRCS ${OS_SRCS} unix/memquery_linux.c)
     set(OS_SRCS ${OS_SRCS} unix/memquery.c)
